+++ phaser/src/menus/battle.ts
@@ -7,7 +7,7 @@ import { Button } from "./button";
 import { Ability, BattleConfig, EFFECT_TYPE, pureCircuits } from "game2-contract";
 import { TestMenu } from "./main";
 import { Subscription } from "rxjs";
-import { AbilityWidget } from "../ability";
+import { AbilityWidget } from "../widgets/ability";
 import { combat_round_logic } from "../battle/logic";
 import { Loader } from "./loader";
 import addScaledImage from "../utils/addScaledImage";
