+++ phaser/src/menus/store.ts
@@ -1,7 +1,7 @@
 import { DeployedGame2API, Game2DerivedState, safeJSONString } from "game2-api";
 import { BattleConfig, pureCircuits } from "game2-contract";
 import { Subscription } from "rxjs";
-import { AbilityWidget } from "../ability";
+import { AbilityWidget } from "../widgets/ability";
 import { fontStyle, GAME_HEIGHT, GAME_WIDTH } from "../main";
 import { Button } from "./button";
 import { Loader } from "./loader";
